# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserSignUpResponse": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"preferCategory": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"email",
					"name",
					"preferCategory"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserSignUpResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/UserSignUpResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSignUpRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"birth": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"detailAddress": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"preferences": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"email",
					"password",
					"name",
					"gender",
					"birth",
					"phoneNumber",
					"preferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserReviewsResponse": {
				"properties": {
					"data": {
						"items": {},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserReviewsResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/UserReviewsResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ReviewResponse": {
				"properties": {
					"reviewId": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"reviewId",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ReviewResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/ReviewResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewCreateRequest": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"body": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"score",
					"body"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewListResponse": {
				"properties": {
					"data": {
						"items": {},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ReviewListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/ReviewListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "umc-nodejs",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/signup": {
			"post": {
				"operationId": "HandleUserSignUp",
				"responses": {
					"200": {
						"description": "회원가입 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserSignUpResponse_"
								}
							}
						}
					},
					"400": {
						"description": "잘못된 요청 데이터 (필수값 누락 등)"
					},
					"409": {
						"description": "이미 존재하는 이메일입니다. (에러코드: U001)"
					}
				},
				"description": "사용자의 이메일, 비밀번호, 이름, 선호 카테고리 등을 입력받아 회원가입을 진행합니다.\r\n성공 시 가입된 사용자의 기본 정보와 선택한 카테고리 목록을 반환합니다.",
				"summary": "사용자 회원가입 API",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserSignUpRequest"
							}
						}
					}
				}
			}
		},
		"/users/{userId}/reviews": {
			"get": {
				"operationId": "HandleListUserReviews",
				"responses": {
					"200": {
						"description": "리뷰 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserReviewsResponse_"
								}
							}
						}
					}
				},
				"description": "특정 사용자가 작성한 가게 리뷰 목록을 페이징 처리하여 최신순으로 조회합니다.",
				"summary": "내 리뷰 목록 조회 API",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/stores/{storeId}/reviews": {
			"post": {
				"operationId": "HandleCreateReview",
				"responses": {
					"200": {
						"description": "리뷰 작성 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ReviewResponse_"
								}
							}
						}
					},
					"400": {
						"description": "잘못된 요청 데이터 (필수값 누락, 점수 범위 초과 등)"
					},
					"404": {
						"description": "요청하신 가게가 존재하지 않습니다."
					}
				},
				"description": "특정 가게(Store)에 대한 사용자의 리뷰를 작성합니다.\r\n성공 시 생성된 리뷰의 고유 ID와 성공 메시지를 반환합니다.",
				"summary": "가게 리뷰 작성 API",
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReviewCreateRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "HandleListStoreReviews",
				"responses": {
					"200": {
						"description": "가게 리뷰 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ReviewListResponse_"
								}
							}
						}
					}
				},
				"description": "특정 가게(Store)에 작성된 리뷰 목록을 페이징 처리하여 조회합니다.",
				"summary": "가게 리뷰 목록 조회 API",
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}